1. 계산결과 저장 프로세스 추가 2. 계산결과 응답 시 계산이력ID 추가 3. 상담상세조회 시 계산결과 세팅 로직 추가 -김웅태

+++ src/main/java/com/xmonster/howtaxing/model/CalculationOwnHouseHistory.java
@@ -21,6 +21,6 @@ public class CalculationOwnHouseHistory implements Serializable {
     private Long ownHouseHistoryId;     // 보유주택이력ID
 
     private Long calcHistoryId;         // 계산이력ID
-    private Integer ownHouseCnt;        // 보유주택수
+    private Long ownHouseCnt;        // 보유주택수
     private Boolean hasOwnHouseDetail;  // 보유주택상세(정보)존재여부
 }
